scripts/requirements.txt
# Dépendances pour scripts/crawler.py et scripts/blog_generator.py (éducation)
firecrawl-py>=1.0.0
google-generativeai>=0.7.0
pydantic>=2.0.0
supabase>=2.0.0
resend>=0.7.0
tenacity>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0
